// apps/academic-records/src/enums/topics.enum.ts
enum Topics {
  AUDIT_LOGS_GET = 'audit_logs.get',
  ACADEMIC_RECORDS_HEALTH_CHECK = 'academic_records.healthCheck',
  STUDENT_GET = 'student.get',
  STUDENT_CREATED = 'student.created',
  NOTIFY_ACADEMIC_RECORDS = 'notify.academic_records',
  ADD_CERTIFICATE = 'add.certificate',
  GET_CERTIFICATE = 'certificate',
  CREATE_APPEAL_ENROLLMENT = 'appeal.created',
  GET_EXAMS = 'get.exams',
  GET_ACADEMIC_RECORDS = 'get.academicRecords',
  GET_STUDENT_RESULTS = 'get.studentResults',
  GRADE_BY_STUDENT = 'grade.by.student',
}

export default Topics;